Fine-tune home page
