+++ packages/client/src/hooks/useCopy.tsx
@@ -1,0 +1,18 @@
+import { useCallback, useState } from 'react';
+
+const useCopy = (): {
+  copiedText: string | null;
+  copyToClipboard: (address: string) => void;
+} => {
+  const [copiedText, setCopiedText] = useState<string | null>(null);
+
+  const copyToClipboard = useCallback((address: string) => {
+    navigator.clipboard.writeText(address);
+    setCopiedText(address);
+    setTimeout(() => setCopiedText(null), 2000);
+  }, []);
+
+  return { copiedText, copyToClipboard };
+};
+
+export default useCopy;
